# Postman/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GestaoColaboradores API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Colaboradores": {
      "get": {
        "tags": [
          "Colaboradores"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Colaborador"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Colaborador"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Colaborador"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Colaboradores"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              }
            }
          }
        }
      }
    },
    "/api/Colaboradores/{id}": {
      "get": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Colaborador"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Colaborador"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Colaboradores"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Unidades": {
      "get": {
        "tags": [
          "Unidades"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Unidade"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Unidade"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Unidade"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Unidades"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              }
            }
          }
        }
      }
    },
    "/api/Unidades/{id}": {
      "get": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unidade"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Unidade"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Unidades"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Colaborador": {
        "type": "object",
        "properties": {
          "idColaborador": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "codigoUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "idUsuario": {
            "type": "integer",
            "format": "int32"
          },
          "unidade": {
            "$ref": "#/components/schemas/Unidade"
          }
        },
        "additionalProperties": false
      },
      "Unidade": {
        "type": "object",
        "properties": {
          "idUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "codigoUnidade": {
            "type": "integer",
            "format": "int32"
          },
          "nomeUnidade": {
            "type": "string",
            "nullable": true
          },
          "unidadeAtiva": {
            "type": "boolean"
          },
          "colaboradores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Colaborador"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}